mcnode: refactor NATConfig to mc/nat
