+++ mc/util.go
@@ -175,7 +175,3 @@ func FilterAddrs(addrs []multiaddr.Multiaddr, predf func(multiaddr.Multiaddr) bo
 	}
 	return res
 }
-
-func GetPublicIP() (string, error) {
-	return "", nil
-}
